# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: WebCMS API
  version: 1.2.0
  description: API documentation for WebCMS v1.2.0
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
paths:
...
  /notifications/{user_id}:
    get:
      summary: Get user notifications
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User notifications
...